overhauled almost all modules, description for more
-new method for enabling modules
-I have defined a is attribute set. It defines lib.optional and its
defined lib.optionals. There are three options as of now it compares;
laptop, desktop, or both.
-these are defined int the flake, accessed via special arg is, and used
like so: is.it.laptop, is.its.desktop, is.it.both (you get the idea)
-is is the attribute set, and it/its determines wether to use a sngle
option or an option set of attributes
-then you just state which host you are checking for.
-my entire configuration is now based around this idea. I changed it to
be this was so I only have to import the autogenerated hardware module
and not set other settings in each hosts' own default files.
-I may be wrong on many of these definitions and I still need to test
these out to see if they even work. just an fyi, but im hoping this will
work